base ghostspawner created
keeps track of when to spawn. spawns random ghosts based on the given number of subtle and unsubtle ghost types needed for that round. for each ghost that is spawned, each of their tamperings will be triggered at equal intervals. all tamperings will finish within a fixed total trigger time