finally got blinking cursor to print properly
added new pixel fonts and a separate font for the cursor. there was a bug with the cursor printing at the very beginning of the last line because monologue.text would automatically print a newline character afterwards, so i needed to remove it. cursor now blinks and prints at the very end of the line rather than at the beginning.