fixed object selection/interaction race condition
when interactable object is selected/clicked on, mouse is locked (in the same position) while monologue is printing. system still thinks it is interacting with the target object's button so pressing spacebar becomes equivalent to interacting with the button. fixed by forcing eventsystem to recognize that nothing/null is selected once printing occurs