fixed autocomplete null bug
firstDemonMatch is set to the very first demon name match inside demonList based on the current input inside the text input. but if there's no match, firstDemonMatch is null. so the original condition for checking tab is invalid bc firstDemonMatch could not be used as a checker unless it is != null