wall object swings, must fix peaks and angle fall
quakeable wall object can swing but at the peak of the swing, the object does not hold for a bit before swinging downward. i thought this would be performed through smoothstep. need to also fix the falling angle so that the object is swinging at its minSwingAngle by the time the swing duration ends